Updated World Generation code

+++ src/omnivoxel/server/client/chunk/worldDataService/noise/Noise3D.java
@@ -1,5 +1,5 @@
 package omnivoxel.server.client.chunk.worldDataService.noise;
 
 public interface Noise3D {
-    double generate(int x, int y, int z);
+    double generate(double x, double y, double z);
 }
